Added a method to get an instructor dashboard url from the backend. Refactored statuses out of the models file.
